+++ src/plugin/markers/index.js
@@ -4,10 +4,13 @@
  * @example wavesurfer.addMarker(regionParams);
  * @property {number} time The time to set the marker at
  * @property {?label} string An optional marker label
+ * @property {?tooltip} string An optional marker tooltip
  * @property {?color} string Background color for marker
  * @property {?position} string "top" or "bottom", defaults to "bottom"
  * @property {?markerElement} element An HTML element to display instead of the default marker image
  * @property {?draggable} boolean Set marker as draggable, defaults to false
+ * @property {?boolean} preventContextMenu Determines whether the context menu
+ * is prevented from being opened, defaults to false
  */
 
 
@@ -66,6 +69,9 @@ export default class MarkersPlugin {
                         this.initPlugin('markers');
                     }
                     return this.markers.add(options);
+                },
+                getMarkers() {
+                    return this.markers;
                 },
                 clearMarkers() {
                     this.markers && this.markers.clear();
@@ -157,9 +163,11 @@ export default class MarkersPlugin {
         let marker = {
             time: params.time,
             label: params.label,
+            tooltip: params.tooltip,
             color: params.color || DEFAULT_FILL_COLOR,
             position: params.position || DEFAULT_POSITION,
-            draggable: !!params.draggable
+            draggable: !!params.draggable,
+            preventContextMenu: !!params.preventContextMenu
         };
 
         marker.el = this._createMarkerElement(marker, params.markerElement);
@@ -168,22 +176,42 @@ export default class MarkersPlugin {
         this.markers.push(marker);
         this._updateMarkerPositions();
 
+        this._registerEvents();
+
         return marker;
     }
 
     /**
      * Remove a marker
      *
-     * @param {number} index Index of the marker to remove
+     * @param {number|Object} indexOrMarker Index of the marker to remove or the marker object itself
      */
-    remove(index) {
+    remove(indexOrMarker) {
+        let index = indexOrMarker;
+        if (isNaN(index)) {
+            index = this.markers.findIndex(marker => marker === indexOrMarker);
+        }
         let marker = this.markers[index];
         if (!marker) {
             return;
         }
+        let label = marker.el.getElementsByClassName("marker-label")[0];
+        if (label) {
+            if (label._onContextMenu) {
+                label.removeEventListener("contextmenu", label._onContextMenu);
+            }
+            if (label._onClick) {
+                label.removeEventListener("click", label._onClick);
+            }
+            if (label._onMouseDown) {
+                label.removeEventListener("mousedown", label._onMouseDown);
+            }
+        }
 
         this.wrapper.removeChild(marker.el);
         this.markers.splice(index, 1);
+
+        this._unregisterEvents();
     }
 
     _createPointerSVG(color, position) {
@@ -216,6 +244,7 @@ export default class MarkersPlugin {
 
     _createMarkerElement(marker, markerElement) {
         let label = marker.label;
+        let tooltip = marker.tooltip;
 
         const el = document.createElement('marker');
         el.className = "wavesurfer-marker";
@@ -250,8 +279,9 @@ export default class MarkersPlugin {
         if ( label ) {
             const labelEl = document.createElement('span');
             labelEl.innerText = label;
+            labelEl.setAttribute('title', tooltip);
             this.style(labelEl, {
-                "font-family": "monospace",
+                "font-family": "inherit",
                 "font-size": "90%"
             });
             labelDiv.appendChild(labelEl);
@@ -262,10 +292,11 @@ export default class MarkersPlugin {
             "align-items": "center",
             cursor: "pointer"
         });
+        labelDiv.classList.add("marker-label");
 
         el.appendChild(labelDiv);
 
-        labelDiv.addEventListener("click", e => {
+        labelDiv._onClick = (e) => {
             e.stopPropagation();
             // Click event is caught when the marker-drop event was dispatched.
             // Drop event was dispatched at this moment, but this.dragging
@@ -275,12 +306,22 @@ export default class MarkersPlugin {
             }
             this.wavesurfer.setCurrentTime(marker.time);
             this.wavesurfer.fireEvent("marker-click", marker, e);
-        });
+        };
+        labelDiv.addEventListener("click", labelDiv._onClick);
+
+        labelDiv._onContextMenu = (e) => {
+            if (marker.preventContextMenu) {
+                e.preventDefault();
+            }
+            this.wavesurfer.fireEvent("marker-contextmenu", marker, e);
+        };
+        labelDiv.addEventListener("contextmenu", labelDiv._onContextMenu);
 
         if (marker.draggable) {
-            labelDiv.addEventListener("mousedown", e => {
+            labelDiv._onMouseDown = () => {
                 this.selectedMarker = marker;
-            });
+            };
+            labelDiv.addEventListener("mousedown", labelDiv._onMouseDown);
         }
         return el;
     }
@@ -358,6 +399,37 @@ export default class MarkersPlugin {
         this.selectedMarker.time = this.wavesurfer.drawer.handleEvent(event) * duration;
         this._updateMarkerPositions();
         this.wavesurfer.fireEvent("marker-drop", this.selectedMarker, event);
+    }
+
+    _registerEvents() {
+        if (!this.markers.find(marker => marker.draggable)) {
+            return;
+        }
+        //we have some draggable markers, check for listeners
+        if (!this.onMouseMove) {
+            this.onMouseMove = (e) => this._onMouseMove(e);
+            window.addEventListener('mousemove', this.onMouseMove);
+        }
+
+        if (!this.onMouseUp) {
+            this.onMouseUp = (e) => this._onMouseUp(e);
+            window.addEventListener("mouseup", this.onMouseUp);
+        }
+    }
+
+    _unregisterEvents() {
+        if (this.markers.find(marker => marker.draggable)) {
+            return;
+        }
+        //we don't have any draggable markers, unregister listeners
+        if (this.onMouseMove) {
+            window.removeEventListener('mousemove', this.onMouseMove);
+            this.onMouseMove = null;
+        }
+        if (this.onMouseUp) {
+            window.removeEventListener("mouseup", this.onMouseUp);
+            this.onMouseUp = null;
+        }
     }
 
     /**